+++ x-pack/plugins/security_solution/public/timelines/components/side_panel/host_details/index.tsx
@@ -82,7 +82,6 @@ export const HostDetailsPanel: React.FC<HostDetailsProps> = React.memo(
           <ExpandableHostDetailsPageLink hostName={hostName} />
           <EuiSpacer size="m" />
           <ExpandableHostDetails contextID={contextID} scopeId={scopeId} hostName={hostName} />
-          <EuiSpacer size="m" />
         </StyledEuiFlyoutBody>
       </>
     ) : (
