+++ x-pack/test/security_solution_cypress/cypress/screens/asset_criticality/flyouts.ts
@@ -29,13 +29,5 @@ export const HOST_DETAILS_FLYOUT_ASSET_CRITICALITY_MODAL_SAVE_BTN = getDataTestS
   'asset-criticality-modal-save-btn'
 );
 
-export const toggleAssetCriticalityAccordion = () => {
-  cy.get(HOST_DETAILS_FLYOUT_ASSET_CRITICALITY_SELECTOR).scrollIntoView();
-  cy.get(HOST_DETAILS_FLYOUT_ASSET_CRITICALITY_SELECTOR).should('be.visible').click();
-};
-
-export const toggleAssetCriticalityModal = () => {
-  toggleAssetCriticalityAccordion();
-
-  cy.get(HOST_DETAILS_FLYOUT_ASSET_CRITICALITY_BUTTON).should('be.visible').click();
-};
+export const OPEN_HOST_FLYOUT_BUTTON = getDataTestSubjectSelector('host-details-button');
+export const OPEN_USER_FLYOUT_BUTTON = getDataTestSubjectSelector('user-details-button');
