+++ x-pack/test/security_solution_cypress/cypress/tasks/asset_criticality/common.ts
@@ -1,0 +1,42 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import {
+  OPEN_HOST_FLYOUT_BUTTON,
+  OPEN_USER_FLYOUT_BUTTON,
+} from '../../screens/asset_criticality/flyouts';
+
+/**
+ * Find the first alert row in the alerts table then click on the host name to open the flyout
+ */
+export const expandFirstAlertHostFlyout = () => {
+  cy.get(OPEN_HOST_FLYOUT_BUTTON).first().click();
+};
+
+/**
+ * Find the first alert row in the alerts table then click on the host name to open the flyout
+ */
+export const expandFirstAlertUserFlyout = () => {
+  cy.get(OPEN_USER_FLYOUT_BUTTON).first().click();
+};
+
+/**
+ * Open the asset criticality accordion
+ */
+export const toggleAssetCriticalityAccordion = () => {
+  cy.get(HOST_DETAILS_FLYOUT_ASSET_CRITICALITY_SELECTOR).scrollIntoView();
+  cy.get(HOST_DETAILS_FLYOUT_ASSET_CRITICALITY_SELECTOR).should('be.visible').click();
+};
+
+/**
+ * Open the asset criticality modal
+ */
+export const toggleAssetCriticalityModal = () => {
+  toggleAssetCriticalityAccordion();
+
+  cy.get(HOST_DETAILS_FLYOUT_ASSET_CRITICALITY_BUTTON).should('be.visible').click();
+};
